changes for connection-model in BE

+++ src/screens/ProfileScreen.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react'
+import React, { useContext, useEffect, useState } from 'react'
 import { useMutation, useQuery } from '@apollo/client'
 import { Button, Container } from '@mui/material';
 import Spinner from '../components/Spinner'
@@ -6,12 +6,15 @@ import UserCard from '../components/profile/UserCard'
 import PartnerCard from '../components/profile/PartnerCard'
 import Search from '../components/profile/Search'
 import RemoveConfirmation from '../components/profile/RemoveConfirmation'
-import { ACCEPT_REQUEST_CONNECTION, CREATE_CHAT, DELETE_CONNECTION, REQUEST_CONNECTION, CREATE_MAP } from '../utils/graphql/mutations'
+import { ACCEPT_REQUEST_CONNECTION, DELETE_CONNECTION, REQUEST_CONNECTION } from '../utils/graphql/mutations'
 import { GET_USER_DETAILS } from '../utils/graphql/queries'
+import {AuthContext} from '../context/auth'
 
 function ProfileScreen({history}) {
     const [partner, setPartner] = useState('')
     const [openConfirmation, setOpenConfirmation] = useState(false);
+
+    const { user } = useContext(AuthContext)
 
     const token = localStorage.getItem('jwtToken')
     
@@ -20,27 +23,36 @@ function ProfileScreen({history}) {
     }
 
     const { loading, data } = useQuery(GET_USER_DETAILS, {
-        onCompleted: () => {
+        onCompleted: (data) => {
             console.log(data)
         },
-        onError: (e) => {
-            console.log({e})
+        onError: (err) => {
+            console.log(JSON.stringify(err, null, 2))
             history.push('/404')
         }     
     })
 
     useEffect(()=>{
-        if(data?.getUserDetails.partner.user){
-            setPartner(data.getUserDetails.partner)
+        if(data?.getUserDetails?.connections[0]?.persons){
+            const filterdPerson = data.getUserDetails.connections[0].persons.filter(person => person.id !== user.id)
+            setPartner({
+                user: filterdPerson[0],
+                status: data.getUserDetails.connections[0].status,
+                requester: data.getUserDetails.connections[0].requester,
+                connectionId: data.getUserDetails.connections[0].id
+            })
         }
-    },[data, setPartner])
+    },[data, setPartner, user.id])
 
     const [requestConnection, {loading: requestLoading}] = useMutation(REQUEST_CONNECTION,{
         refetchQueries: [
             {
                 query: GET_USER_DETAILS
             }
-          ]
+          ],
+        onError: (err) => {
+            console.log(JSON.stringify(err, null, 2))
+        }  
     })
 
     const [acceptRequestConnection, {loading: acceptLoading}] = useMutation(ACCEPT_REQUEST_CONNECTION,{
@@ -48,7 +60,10 @@ function ProfileScreen({history}) {
             {
                 query: GET_USER_DETAILS
             }
-          ]
+        ],
+        onError: (err) => {
+            console.log(JSON.stringify(err, null, 2))
+        }  
     })
 
     const [deleteConnection, {loading: deleteLoading}] = useMutation(DELETE_CONNECTION,{
@@ -56,23 +71,10 @@ function ProfileScreen({history}) {
             {
                 query: GET_USER_DETAILS
             }
-          ]
-    })
-
-    const [createChat, {loading: chatLoading}] = useMutation(CREATE_CHAT,{
-        refetchQueries: [
-            {
-                query: GET_USER_DETAILS
-            }
-          ]
-    })
-
-    const [createMap, {loading: mapLoading}] = useMutation(CREATE_MAP,{
-        refetchQueries: [
-            {
-                query: GET_USER_DETAILS
-            }
-          ]
+        ],
+        onError: (err) => {
+            console.log(JSON.stringify(err, null, 2))
+        }  
     })
 
     const handleRequest = (e) => {
@@ -85,28 +87,27 @@ function ProfileScreen({history}) {
     const handleAccept = (e) => {
         e.preventDefault()
         acceptRequestConnection({variables: {
-            partnerId: partner.user.id} 
+            connectionId: e.target.id} 
         })
-        createChat({variables: {
-            partnerId: partner.user.id} 
-        })
-        createMap({variables: {
-            partnerId: partner.user.id} 
-        })
+        console.log(e.target.id)
     }
-
+    
     const handleRemove = (e) => {
         e.preventDefault()
-        deleteConnection({variables: {
-            partnerId: partner.user.id} 
-        })
-        setPartner('')
+        if(data.getUserDetails?.email !== process.env.REACT_APP_DEMO_MAIL){
+            window.alert("This is just a demo! This function is not working here")
+        } else {
+            deleteConnection({variables: {
+                partnerId: partner.user.id} 
+            })
+            setPartner('')   
+        }
     }
 
     return (
         <div>
-            {mapLoading ||requestLoading || acceptLoading || deleteLoading || chatLoading || loading ? (
-                <Spinner loading={mapLoading ||requestLoading || acceptLoading || deleteLoading || chatLoading || loading} />
+            {requestLoading || acceptLoading || deleteLoading || loading ? (
+                <Spinner loading={requestLoading || acceptLoading || deleteLoading || loading} />
             ) : (
                 <Container>
                     <h3>Your Profile</h3>
@@ -115,12 +116,12 @@ function ProfileScreen({history}) {
                         <Search setPartner={setPartner} /> 
                         : 
                         <>
-                            <h3>Your Connection</h3>
+                            <h3>Your Connection {partner.status === "pending" && "is pending"}</h3>
                             <PartnerCard user={partner.user} />
-                            {partner.status === "pending" ? 
-                            <Button variant="contained" onClick={handleAccept}>Accept</Button>
+                            {(partner.status === "pending" && partner.requester !== user.id) ? 
+                            <Button style={{marginTop: "2rem", marginRight: "2rem"}} variant="contained" id={partner.connectionId} onClick={handleAccept}>Accept</Button>
                             : partner.status === "none" ?
-                            <Button variant="contained" onClick={handleRequest}>Request</Button>
+                            <Button style={{marginTop: "2rem", marginRight: "2rem"}} variant="contained" onClick={handleRequest}>Request</Button>
                             :
                             ""
                             }
